refactor(agent): extract calculateReplaceRange.
