+++ clients/tabby-agent/tests/bad_cases/2-typescript.ts
@@ -1,0 +1,3 @@
+function clamp(n: number, max: number, min: number): number {
+  return Math.max(Math.min(⏩⏭n, max), min⏮⏪);
+}
